#30 - @josigama @viniciusgama - add new file for services and get port from env variable
