+++ src/app.js
@@ -1,6 +1,7 @@
 var restify = require('restify');
 var database = require('./db.connection');
-var Place = require('./place.model');
+var services = require('./services');
+var port = process.env.VAMOS_JUNTAS_PORT || 8080;
 
 module.exports = {
   start: function () {
@@ -10,33 +11,11 @@ module.exports = {
 
     server.use(restify.queryParser());
 
-    server.get('/risks-around', function (request, response) {
-      if (Object.keys(request.params).length === 0) {
-        return response.send(300, 'Invalid params.');
-      }
+    server.get('/risks-around', services.risksAround);
 
-      var maxDistance = 100;
-      var coords = [];
-      coords[0] = request.params.longitude;
-      coords[1] = request.params.latitude;
-
-      Place.find({
-        loc: {
-          $near: coords,
-          $maxDistance: maxDistance
-        }
-      }).exec(function (err, places) {
-        if (err) {
-          return console.error(err);
-        }
-        console.log(places);
-        return response.send(places);
-      });
-    });
-
-    server.listen(8080, function() {
+    server.listen(port, function() {
       database.open();
-    	console.log('Hello World! :D');
+    	console.log('Vamos Juntas Start :D');
     });
   }
 };
